修改部分文件名称 创建时间和更新时间字段名称修改

+++ bin/controller/expense_controller.js
@@ -5,8 +5,8 @@
 
 var express = require('express');
 var router = express.Router();
-var expense = require('../lib/expenseItem_service.js');
-var etils = require('../lib/etils.js');
+var expense = require('../lib/expense_service.js');
+var tils = require('../lib/tils.js');
 router.get('/', function (req, res) {
     var userid = req.session.user?req.session.user.uid:null;
     if(userid){
@@ -34,7 +34,7 @@ router.post('/', function (req, res) {
     var detail = req.body.detail;
     var money = req.body.expense;
     var expenseTime = req.body.expenseTime;
-    var eid = etils.newUuid();
+    var eid = tils.newUuid();
     var userid = req.session.user?req.session.user.uid:null;
     if(userid){
         expense.addOne(detail, money, eid,expenseTime,userid)
@@ -62,35 +62,51 @@ router.put('/:eid', function (req, res) {
     var detail = req.body.detail?req.body.detail:null;
     var money = req.body.expense?req.body.expense:null;
     var expenseTime = req.body.expenseTime?req.body.expenseTime:null;
-    expense.updateOne(eid,detail,money,expenseTime)
-        .then((result)=>{
-            res.send({
-                data: result,
-                status: 0
-            });
-        })
-        .catch((err) => {
-            res.send({
-                status: -1
-            });
-            console.log(err);
-        })
+    var userid = req.session.user?req.session.user.uid:null;
+    if(userid) {
+        expense.updateOne(eid, detail, money, expenseTime)
+            .then((result) => {
+                res.send({
+                    data: result,
+                    status: 0
+                });
+            })
+            .catch((err) => {
+                res.send({
+                    status: -1
+                });
+                console.log(err);
+            })
+    }else{
+        res.send({
+            status: -1,
+            message:"请先登录再添加消费记录！"
+        });
+    }
 });
 router.delete('/:eid', function (req, res) {
     var eid = req.params.eid;
-    expense.deleteOne(eid)
-        .then((result)=>{
-            res.send({
-                data: result,
-                status: 0
-            });
-        })
-        .catch((err) => {
-            res.send({
-                status: -1
-            });
-            console.log(err);
-        })
+    var userid = req.session.user?req.session.user.uid:null;
+    if(userid) {
+        expense.deleteOne(eid)
+            .then((result) => {
+                res.send({
+                    data: result,
+                    status: 0
+                });
+            })
+            .catch((err) => {
+                res.send({
+                    status: -1
+                });
+                console.log(err);
+            })
+    }else{
+        res.send({
+            status: -1,
+            message:"请先登录再添加消费记录！"
+        });
+    }
 });
 
 module.exports = router;
